Move LoginMenuScreen and InputDialog to menu package
