+++ core/src/main/kotlin/no/ntnu/beardblaster/menu/InputDialog.kt
