+++ core/src/main/kotlin/no/ntnu/beardblaster/menu/MenuPresenter.kt
@@ -9,7 +9,6 @@ import no.ntnu.beardblaster.leaderboard.LeaderBoardScreen
 import no.ntnu.beardblaster.lobby.JoinLobbyScreen
 import no.ntnu.beardblaster.lobby.LobbyScreen
 import no.ntnu.beardblaster.tutorial.TutorialScreen
-import no.ntnu.beardblaster.user.LoginMenuScreen
 import no.ntnu.beardblaster.user.UserAuth
 import no.ntnu.beardblaster.user.UserData
 import java.util.*
